// types/SecurityGroupRole.ts
export interface SecurityGroupRole {
    SECURITY_GROUP_ROLE_UUID: string;
    SECURITY_RESOURCE_UUID?: string;
    SECURITY_APP_EAI_NBR: number;
    SECURITY_GROUP_UUID: string;
    SECURITY_ROLE_UUID: string;
    LAST_UPD_DT_TM: string;
    LAST_UPD_USER: string;
    CHANGE_FLAG: string;
    GROUP_NAME?: any;
}